Move vmi transition histograms to perfscale packages
