+++ pkg/monitoring/perfscale/perfscale_suite_test.go
@@ -1,0 +1,13 @@
+package perfscale_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestPerfscale(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Perfscale Suite")
+}
